add utility classes and data models for wallpaper management

+++ app/src/main/java/io/droidevs/wallpaper/domain/usecases/data/collection/GetCollectionByIdUseCase.kt
@@ -12,7 +12,7 @@ class GetCollectionByIdUseCase(
     private val repository: CollectionRepository,
     private val dispatchers: AppDispatchers
 ) {
-    operator fun invoke(id: Int): Flow<Result<Collection?, DataError>> {
+    operator fun invoke(id: Long): Flow<Result<Collection?, DataError>> {
         return repository.getCollectionById(id).flowOn(
             dispatchers.io
         )
